[task 1, stage 6] декомпозиция карточки
