+++ ReactTasks/task-1/stage-6/src/pages/TodoLists/components/Card/Card.module.scss
@@ -33,13 +33,15 @@
     width: rem(16);
     height: rem(16);
   }
+  &__header{
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-end;
+    gap: rem(20);
+    width: 100%;
+  }
   &__control {
     display: flex;
     gap: rem(5);
-  }
-  &__form {
-    padding: 0;
-    border: none;
-    box-shadow: none;
   }
 }
