+++ ReactTasks/task-1/stage-6/src/pages/TodoLists/components/FormEditingCard/FormEditingCard.module.scss
@@ -1,0 +1,5 @@
+.formEditingCardForm {
+  padding: 0;
+  border: none;
+  box-shadow: none;
+}
